feat: Introduce main API and tests
